Finalize all ViewModels and logic (most likely)

+++ Wintertale.Client/ViewModels/Dashboard/DashboardViewModel.cs
@@ -2,6 +2,7 @@
 using CommunityToolkit.Mvvm.Input;
 using System.Collections.ObjectModel;
 using System.Diagnostics;
+using System.Text.RegularExpressions;
 using System.Threading.Tasks;
 using Wintertale.Client.Common.DTOs.Requests.Friends;
 using Wintertale.Client.Common.DTOs.Responses.Auth;
@@ -25,6 +26,9 @@ namespace Wintertale.Client.ViewModels.Dashboard {
         [ObservableProperty]
         private ObservableCollection<FriendResponse> friends;
 
+        [ObservableProperty]
+        private ObservableCollection<FriendResponse> pendingFriends;
+
         public async void ApplyQueryAttributes(IDictionary<string, object> query) {
             currentUser = query["CurrentUser"] as LoginResponse;
 
@@ -36,11 +40,16 @@ namespace Wintertale.Client.ViewModels.Dashboard {
             }
 
             Friends = new(await service.GetFriendListAsync());
+            PendingFriends = new(await service.GetPendingFriendsReceivedAsync());
         }
 
         [RelayCommand]
         private async Task OpenFriends() {
+            var navigationParameters = new Dictionary<string, object> {
+                { "CurrentUser", currentUser! },
+            };
 
+            await Shell.Current.GoToAsync("FriendRequestsPage", navigationParameters);
         }
 
         [RelayCommand]
@@ -48,14 +57,72 @@ namespace Wintertale.Client.ViewModels.Dashboard {
             await Shell.Current.GoToAsync("..");
         }
 
+        [ObservableProperty]
+        private string friendPhone;
+
+        [RelayCommand]
+        private async Task CreateFriend() {
+            string formattedPhone = Regex.Replace(FriendPhone, @"\D", "");
+            if (!Regex.IsMatch(formattedPhone, @"^7\d{10}$")) {
+                await Shell.Current.DisplayAlertAsync("Ошибка валидации", formattedPhone, "ОК");
+                return;
+            }
+
+            var request = new CreateFriendRequest {
+                phone = formattedPhone
+            };
+
+            try {
+                var response = await service.CreateFriendAsync(request);
+                await Shell.Current.DisplayAlertAsync("Успех", $"Приглашение отправлено пользователю: {formattedPhone}", "ОК");
+            } catch (Exception ex) {
+                await Shell.Current.DisplayAlertAsync("Ошибка приглашения", ex.Message, "ОК");
+            }
+        }
+
+        [RelayCommand]
+        private async Task UpdateFriend(FriendResponse friend) {
+            var request = new UpdateFriendRequest {
+                phone = friend.phone,
+                status = 1
+            };
+
+            try {
+                await service.UpdateFriendAsync(request);
+                PendingFriends.Remove(friend);
+                Friends.Add(friend);
+            } catch (Exception ex) {
+                await Shell.Current.DisplayAlertAsync("Ошибка принятия запроса", ex.Message, "ОК");
+            }
+        }
+
+        [RelayCommand]
+        private async Task RemovePendingFriend(FriendResponse friend) {
+            var request = new RemoveFriendRequest {
+                phone = friend.phone
+            };
+
+            try {
+                await service.RemoveFriendAsync(request);
+                PendingFriends.Remove(friend);
+            } catch (Exception ex) {
+                await Shell.Current.DisplayAlertAsync("Ошибка удаления", ex.Message, "ОК");
+            }
+        }
+
         [RelayCommand]
         private async Task RemoveFriend(FriendResponse friend) {
             var request = new RemoveFriendRequest {
                 phone = friend.phone
             };
-
-            await service.RemoveFriendAsync(request);
-            Friends.Remove(friend);
+            if (await Shell.Current.DisplayAlertAsync("Удалить из друзей", $"Вы действительно хотите удалить {friend.fname} из списка друзей?", "Удалить друга", "Отменить")) {
+                try {
+                    await service.RemoveFriendAsync(request);
+                    Friends.Remove(friend);
+                } catch (Exception ex) {
+                    await Shell.Current.DisplayAlertAsync("Ошибка удаления", ex.Message, "ОК");
+                }
+            }
         }
 
         [ObservableProperty]
